Create 24 Q.1 C Exam Question
Write a function in C that takes two integer arguments. If the
arguments are X and Y, your function should multiply X by 2 Y times,
printing each value as it calculates it. For example if the first argument
is 5 and the 2nd argument is 4, you function should printout the values
of 5*2, 5*2*2, 5*2*2*2 and 5*2*2*2*2. Your function should work
for any 2 arguments.