Create 24 Q.2
Write a function in C that takes one argument of type float. The
function should repeatedly multiply the argument by 1.5 and then
subtract 5. It should keep doing this until the value is greater than 100.
Your function should then return the number of times that it performed
the multiplication and subtraction. For example if the function was
passed the argument 10.0 the next value would be 25 ((20*1.5)-5).
Your function should work for any value passed to it.